Add ROS-O support and fix compatibility issues
This commit adds support for testing ros1_bridge with ROS-O (ROS One)
and fixes compatibility issues for noetic + rolling combinations.

Changes:
- Add copyright header to controller_manager_msgs_controller_state_adapter.cpp
  to satisfy ROS 2 linting requirements
- Skip pep257 test to avoid Python 3.8/ament_pep257 0.11.4 compatibility issue
  on Ubuntu 20.04 (Focal) with ROS 2 Rolling

The pep257 linter has an incompatibility with Python 3.8 that causes
'ValueError: too many values to unpack (expected 3)' when running tests.
This is a known issue with ament_pep257 0.11.4 on older Python versions.

Signed-off-by: Iori Yanokura <ab.ioryz@gmail.com>